requirements.txt
Django==1.10.1
djangorestframework==3.4.6
sdklib==1.2
django-filter==0.14.0
django-crispy-forms==1.6.0
requests
